+++ backend/team6/src/main/java/programmers/team6/domain/admin/service/VacationStatisticsService.java
@@ -7,6 +7,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import programmers.team6.domain.admin.dto.VacationStatisticsRequest;
+import programmers.team6.domain.admin.support.MemberReader;
+import programmers.team6.domain.admin.support.Members;
+import programmers.team6.domain.admin.support.VacationInfoLogReader;
+import programmers.team6.domain.admin.support.VacationInfoLogs;
+import programmers.team6.domain.admin.support.VacationRequests;
+import programmers.team6.domain.admin.support.VacationRequestsReader;
 import programmers.team6.domain.admin.utils.mapper.VacationStatisticsMapper;
 import programmers.team6.domain.member.repository.MemberRepository;
 import programmers.team6.domain.vacation.dto.VacationMonthlyStatisticsResponse;
