+++ backend/team6/src/main/java/programmers/team6/domain/admin/support/VacationRequests.java
@@ -1,4 +1,4 @@
-package programmers.team6.domain.admin.service;
+package programmers.team6.domain.admin.support;
 
 import java.util.Collections;
 import java.util.List;
