+++ backend/team6/src/main/java/programmers/team6/domain/vacation/service/VacationInfoService.java
@@ -15,7 +15,8 @@ import programmers.team6.domain.vacation.entity.VacationInfoLog;
 import programmers.team6.domain.vacation.repository.VacationInfoRepository;
 import programmers.team6.domain.vacation.rule.VacationGrantRule;
 import programmers.team6.domain.vacation.rule.VacationGrantRuleFinder;
-import programmers.team6.domain.vacation.rule.VacationInfos;
+import programmers.team6.domain.vacation.support.VacationInfos;
+import programmers.team6.domain.vacation.support.VacationInfoLogPublisher;
 import programmers.team6.global.exception.code.BadRequestErrorCode;
 import programmers.team6.global.exception.code.ConflictErrorCode;
 import programmers.team6.global.exception.customException.BadRequestException;
