+++ backend/team6/src/main/java/programmers/team6/domain/vacation/support/VacationInfos.java
@@ -1,4 +1,4 @@
-package programmers.team6.domain.vacation.rule;
+package programmers.team6.domain.vacation.support;
 
 import java.util.Collections;
 import java.util.List;
