+++ backend/team6/src/test/java/programmers/team6/domain/vacation/service/ApprovalStepServiceTests.java
@@ -27,6 +27,7 @@ import programmers.team6.domain.vacation.enums.ApprovalStatus;
 import programmers.team6.domain.vacation.enums.VacationRequestStatus;
 import programmers.team6.domain.vacation.repository.ApprovalStepRepository;
 import programmers.team6.domain.vacation.repository.VacationInfoRepository;
+import programmers.team6.domain.vacation.support.VacationInfoLogPublisher;
 import programmers.team6.global.exception.customException.BadRequestException;
 import programmers.team6.global.exception.customException.NotFoundException;
 import programmers.team6.support.MemberMother;
