+++ backend/team6/src/test/java/programmers/team6/domain/vacation/service/VacationInfoServiceTest.java
@@ -21,6 +21,7 @@ import programmers.team6.domain.vacation.entity.VacationInfoLog;
 import programmers.team6.domain.vacation.repository.VacationInfoRepository;
 import programmers.team6.domain.vacation.rule.VacationGrantRule;
 import programmers.team6.domain.vacation.rule.VacationGrantRuleFinder;
+import programmers.team6.domain.vacation.support.VacationInfoLogPublisher;
 import programmers.team6.global.exception.customException.BadRequestException;
 import programmers.team6.global.exception.customException.ConflictException;
