+++ backend/team6/src/test/java/programmers/team6/mock/MemberReaderFake.java
@@ -8,8 +8,8 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 
 import programmers.team6.domain.admin.dto.VacationStatisticsRequest;
-import programmers.team6.domain.admin.service.MemberReader;
-import programmers.team6.domain.admin.service.Members;
+import programmers.team6.domain.admin.support.MemberReader;
+import programmers.team6.domain.admin.support.Members;
 import programmers.team6.domain.member.entity.Member;
 
 public class MemberReaderFake extends MemberReader {
