+++ backend/team6/src/test/java/programmers/team6/mock/VacationInfoLogReaderFake.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Map;
 
 import programmers.team6.domain.admin.dto.VacationStatisticsRequest;
-import programmers.team6.domain.admin.service.VacationInfoLogReader;
-import programmers.team6.domain.admin.service.VacationInfoLogs;
+import programmers.team6.domain.admin.support.VacationInfoLogReader;
+import programmers.team6.domain.admin.support.VacationInfoLogs;
 import programmers.team6.domain.vacation.entity.VacationInfoLog;
 
 public class VacationInfoLogReaderFake extends VacationInfoLogReader {
